refactor(apollo-vertex): remove unused Nextjs pages now rendered by Tanstack Router
