--- apps/apollo-vertex/app/(preview)/layout.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export default function PreviewLayout({ children }: { children: ReactNode }) {
-  return <div className="fixed inset-0 z-50 bg-background">{children}</div>;
-}
