--- apps/apollo-vertex/app/(preview)/preview/shell-minimal/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { LoanQcDashboard } from "./loan-qc-dashboard";
-
-export default function ShellMinimalPreviewPage() {
-  return <LoanQcDashboard />;
-}
