--- apps/apollo-vertex/app/(preview)/preview/shell/dashboard/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { InvoiceDashboard } from "../invoice-dashboard";
-
-export default function DashboardPage() {
-  return <InvoiceDashboard />;
-}
