--- apps/apollo-vertex/app/(preview)/preview/shell/layout.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { ShellTemplate } from "@/templates/shell/ShellTemplateDynamic";
-
-export default function ShellLayout() {
-  return <ShellTemplate />;
-}
