--- apps/apollo-vertex/app/(preview)/preview/shell/page.tsx
@@ -1,12 +1,0 @@
-export default function ShellPreviewPage() {
-  return (
-    <div className="flex items-center justify-center h-full">
-      <div className="text-center space-y-2">
-        <h1 className="text-2xl font-semibold">Select a page</h1>
-        <p className="text-muted-foreground">
-          Choose a page from the sidebar to get started.
-        </p>
-      </div>
-    </div>
-  );
-}
