--- apps/apollo-vertex/app/(preview)/preview/shell/projects/page.tsx
@@ -1,12 +1,0 @@
-export default function ProjectsPage() {
-  return (
-    <div className="flex items-center justify-center h-full">
-      <div className="text-center space-y-2">
-        <h1 className="text-2xl font-semibold">Projects</h1>
-        <p className="text-muted-foreground">
-          This page is under construction.
-        </p>
-      </div>
-    </div>
-  );
-}
