--- apps/apollo-vertex/app/(preview)/preview/shell/settings/page.tsx
@@ -1,12 +1,0 @@
-export default function SettingsPage() {
-  return (
-    <div className="flex items-center justify-center h-full">
-      <div className="text-center space-y-2">
-        <h1 className="text-2xl font-semibold">Settings</h1>
-        <p className="text-muted-foreground">
-          This page is under construction.
-        </p>
-      </div>
-    </div>
-  );
-}
